refactor: DataProvider 범위를 CartPage로 제한
