+++ src/App.tsx
@@ -4,21 +4,18 @@ import Layout from './components/Layout/Layout';
 import CartPage from './pages/CartPage';
 import GlobalStyle from './styles/globalStyle';
 import OrderConfirmPage from './pages/OrderConfirmPage';
-import { DataProvider } from './context/DataContext';
 
 function App() {
   return (
     <>
       <Global styles={GlobalStyle} />
       <Layout>
-        <DataProvider>
-          <BrowserRouter basename="/react-shopping-cart">
-            <Routes>
-              <Route path="/" element={<CartPage />} />
-              <Route path="/orderConfirm" element={<OrderConfirmPage />} />
-            </Routes>
-          </BrowserRouter>
-        </DataProvider>
+        <BrowserRouter basename="/react-shopping-cart">
+          <Routes>
+            <Route path="/" element={<CartPage />} />
+            <Route path="/orderConfirm" element={<OrderConfirmPage />} />
+          </Routes>
+        </BrowserRouter>
       </Layout>
     </>
   );
